requirements/base.txt
#Django==1.8.6
#bpython==0.13
django-braces==1.8.1
django-model-utils==2.3.1
logutils==0.3.3
psycopg2==2.6.1
static==1.1.1
dj-database-url==0.3.0
dj-static==0.0.6
#django-debug-toolbar==1.2.1
